refactor: EpigramDetailCard 위젯 분리 (#51)
- views/epigram-detail의 인라인 EpigramCard·RuledLines·Reference·룰 라인 상수를
  widgets/epigram-detail-card로 추출
- widgets/epigram-card(피드용)와의 이름 충돌 해소
- EpigramDetailScreen은 페이지 chrome(BackButton·로딩/에러 상태)만 남김
