+++ src/widgets/epigram-detail-card/index.ts
@@ -1,0 +1,1 @@
+export { EpigramDetailCard } from "./ui/EpigramDetailCard";
